EMail Struct organized and added an automatic Mail-Sender after every Payment Curently working on a possible cronjob to send Weekly Mails for Kantiniers, but im struggling with the golang return values in dbP. It seems like an easy fix for people who are used to go.
