+++ server/data/SMTPServerStruct.go
@@ -1,0 +1,15 @@
+package data
+
+import "crypto/tls"
+
+// SMTPServer ...
+type SMTPServer struct {
+	Host      string
+	Port      string
+	TLSConfig *tls.Config
+}
+
+// ServerName ...
+func (s *SMTPServer) ServerName() string {
+	return s.Host + ":" + s.Port
+}
