+++ server/database/bookingsFunctions.go
@@ -3,6 +3,7 @@ package database
 import (
 	"fmt"
 	"log"
+	"os"
 	"strconv"
 	"time"
 
@@ -166,10 +167,37 @@ func Pay(payment data.Payment) bool {
 	err = tx.Commit()
 	HandleDatabaseError(err)
 
+	// Set new User.Balance based on the payment amount.
 	newBalance := payment.User.Balance - payment.Balance
 	query := fmt.Sprintf("UPDATE users SET balance = %.2f WHERE id = %d;", newBalance, payment.User.ID)
 	_, err = db.Query(query)
 	HandleDatabaseError(err)
+
+	if !true { // ToDo: Check if sending an EMail after every payment is activated in the settings.
+		return err == nil
+	}
+	// Send an E-Mail to the User, with all payment information included.
+	receiver := payment.User.Email
+	if len(receiver) == 0 {
+		log.Println("Mail to " + payment.User.FirstName + " (" + payment.User.BierName + ") " + payment.User.LastName + " failed. Because no E-Mail is configured for the user.")
+		return err == nil
+	}
+	receiverCC := ""                        //Normally noone needs to receive this mail in CC
+	receiverBCC := ""                       //But it could be possible to set a BCC receiver to the one in charge with the Bank Account
+	sender := os.Getenv("AVHBS_EMAIL_USER") //ToDo: Change to Database Setting E-Mail
+
+	//From here on think of internationalisation or Database Settings for the texts.
+	subject := "New Payment received"
+	message := "<p>Hello dear " + payment.User.FirstName + " (" + payment.User.BierName + ") " + payment.User.LastName + ",<br>" +
+		"We received a payment from you.</p>" +
+		"<table><tbody>" +
+		"<tr><td>User Balance:</td><td>" + fmt.Sprintf("%.2f", (payment.User.Balance*-1)) + "</td></tr>" +
+		"<tr><td>Amount Payed:</td><td>" + fmt.Sprintf("%.2f", payment.Balance) + "</td></tr>" +
+		"<tr><td>New User Balance:</td><td>" + fmt.Sprintf("%.2f", (newBalance*-1)) + "</td></tr>" +
+		"<tr><td>Payment method:</td><td>" + payment.PaymentMethod + "</td></tr>" +
+		"</tbody></table><br>" +
+		"<p>Thank you for your payment on " + timeStamp + "</p>"
+	data.EmailController(receiver, receiverCC, receiverBCC, sender, subject, message)
 	return err == nil
 }
